Move date time crate arg enum conversion into commands module
