+++ sea-orm-cli/src/cli.rs
@@ -1,5 +1,5 @@
-use clap::{ArgEnum, ArgGroup, Parser, Subcommand};
-use sea_orm_codegen::DateTimeCrate as CodegenDateTimeCrate;
+use clap::{ArgGroup, Parser, Subcommand};
+use crate::commands::DateTimeCrate;
 
 #[derive(Parser, Debug)]
 #[clap(version)]
@@ -167,19 +167,4 @@ pub enum GenerateSubcommands {
         )]
         date_time_crate: DateTimeCrate,
     },
-}
-
-#[derive(ArgEnum, Copy, Clone, Debug, PartialEq)]
-pub enum DateTimeCrate {
-    Chrono,
-    Time,
-}
-
-impl From<DateTimeCrate> for CodegenDateTimeCrate {
-    fn from(date_time_crate: DateTimeCrate) -> CodegenDateTimeCrate {
-        match date_time_crate {
-            DateTimeCrate::Chrono => CodegenDateTimeCrate::Chrono,
-            DateTimeCrate::Time => CodegenDateTimeCrate::Time,
-        }
-    }
 }
